requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
app==0.0.1
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
boto3==1.28.58
botocore==1.31.58
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
databases==0.4.2
datefinder==0.7.3
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.103.1
firebase-admin==6.1.0
flake8==6.1.0
forex-python==1.8
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googlemaps==4.10.0
gotrue==2.4.2
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==21.2.0
h11==0.14.0
h2==3.2.0
hashids==1.3.1
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
hyperframe==5.2.0
idna==3.4
iniconfig==2.0.0
inotify==0.2.10
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
maxminddb==2.5.2
mccabe==0.7.0
motor==3.2.0
msg91-otp==0.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nose==1.3.7
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.4
passlib==1.7.4
pathspec==0.11.2
phonenumbers==8.13.24
platformdirs==3.10.0
pluggy==1.3.0
postgrest==0.16.4
proto-plus==1.22.3
protobuf==4.23.3
psycopg2==2.9.9
psycopg2-binary==2.9.7
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
pymongo==4.4.1
pyparsing==3.1.0
pypng==0.20220715.0
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
qrcode==7.4.2
ratelim==0.1.6
razorpay==1.3.0
realtime==1.0.4
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.7.0
sendgrid==6.10.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starkbank-ecdsa==2.2.0
starlette==0.27.0
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
Telethon==1.34.0
timer==0.2.2
tomli==2.0.1
tqdm==4.66.4
twilio==8.10.2
typing_extensions==4.11.0
tzdata==2023.3
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.17
user-agent2==2021.12.11
user-agents==2.2.0
user_agent==0.1.10
utils==1.0.1
uvicorn==0.23.2
websockets==12.0
yarl==1.9.3
